+++ image_recognition_util/src/image_recognition_util/__init__.py
@@ -1,1 +1,1 @@
-from .colour_extractor import ColourExtractor
+from .color_extractor import ColorExtractor
